airbyte-ci: factor git checkout
