+++ airbyte-ci/connectors/pipelines/pipelines/helpers/git.py
@@ -3,12 +3,14 @@
 #
 
 import functools
-from typing import List, Optional, Set
+from typing import List, Set
 
 import git
-from dagger import Client, Connection, Container
+from dagger import Connection
 from github import PullRequest
-from pipelines.helpers.utils import AIRBYTE_REPO_URL, DAGGER_CONFIG, DIFF_FILTER
+
+from pipelines.dagger.containers.git import checked_out_git_container
+from pipelines.helpers.utils import DAGGER_CONFIG, DIFF_FILTER
 
 
 def get_current_git_revision() -> str:  # noqa D103
@@ -78,37 +80,6 @@ async def get_modified_files_in_commit(current_git_branch: str, current_git_revi
 def get_modified_files_in_pull_request(pull_request: PullRequest) -> List[str]:
     """Retrieve the list of modified files in a pull request."""
     return [f.filename for f in pull_request.get_files()]
-
-
-async def checked_out_git_container(
-    dagger_client: Client,
-    current_git_branch: str,
-    current_git_revision: str,
-    diffed_branch: Optional[str] = None,
-) -> Container:
-    current_git_branch = current_git_branch.removeprefix("origin/")
-    diffed_branch = current_git_branch if diffed_branch is None else diffed_branch.removeprefix("origin/")
-    return await (
-        dagger_client.container()
-        .from_("alpine/git:latest")
-        .with_workdir("/repo")
-        .with_exec(["init"])
-        .with_env_variable("CACHEBUSTER", current_git_revision)
-        .with_exec(
-            [
-                "remote",
-                "add",
-                "--fetch",
-                "--track",
-                current_git_branch,
-                "--track",
-                diffed_branch if diffed_branch is not None else current_git_branch,
-                "origin",
-                AIRBYTE_REPO_URL,
-            ]
-        )
-        .with_exec(["checkout", "-t", f"origin/{current_git_branch}"])
-    )
 
 
 @functools.cache
